requirements.txt
# Pinned for reproducible installs. Bump versions after smoke-testing.
google-generativeai==0.8.4
streamlit==1.40.2
python-dotenv==1.0.1
pymupdf==1.24.14
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1
python-pptx==1.0.2
